content/html/content/test/test_bug377624.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=377624
-->
<head>
  <title>Test for Bug 377624</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=377624">Mozilla Bug 377624</a>
<p id="display"></p>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 377624 **/

var input = document.createElement('input');
ok("accept" in input, "'accept' is a valid input property");

</script>
</pre>
</body>
</html>
